requirements.txt
pandas==1.5.3
openpyxl>=3.1.0
matplotlib==3.6.2
twilio==7.16.0
fpdf2==2.7.4
python-dotenv==0.21.0
requests==2.28.2
schedule==1.2.0